moved collection index functionality to seprate component
