--- ui/src/components/Collection/index.jsx
@@ -1,3 +1,0 @@
-import CollectionListItem from './CollectionListItem';
-
-export { CollectionListItem };
