+++ ui/src/components/CollectionIndex/CollectionIndex.scss
@@ -1,6 +1,6 @@
 @import "src/app/variables.scss";
 
-.CollectionListItem {
+.CollectionIndexItem {
     padding: $aleph-grid-size*2.5 0;
     border-bottom: 1px solid $aleph-border-color;
