+++ ui/src/components/CollectionIndex/CollectionIndexSearch.scss
