+++ ui/src/components/CollectionIndex/index.jsx
@@ -1,0 +1,3 @@
+import CollectionIndexItem from './CollectionIndexItem';
+
+export { CollectionIndexItem };
