+++ ui/src/screens/GroupScreen/GroupScreen.jsx
@@ -1,38 +1,32 @@
 import React, { Component } from 'react';
 import {
-  defineMessages, injectIntl, FormattedMessage,
+  // defineMessages,
+  injectIntl,
+  FormattedMessage,
 } from 'react-intl';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
-import { Waypoint } from 'react-waypoint';
 
 import Query from 'src/app/Query';
 import Dashboard from 'src/components/Dashboard/Dashboard';
-import { SectionLoading, ErrorSection } from 'src/components/common';
 import Screen from 'src/components/Screen/Screen';
-import CollectionListItem from 'src/components/Collection/CollectionListItem';
+import CollectionIndex from 'src/components/CollectionIndex/CollectionIndex';
 import LoadingScreen from 'src/components/Screen/LoadingScreen';
 import { queryCollections, fetchGroups } from 'src/actions';
-import { selectCollectionsResult, selectGroups, selectRole } from 'src/selectors';
+import { selectGroups, selectRole } from 'src/selectors';
 
 import './GroupScreen.scss';
 
 
-const messages = defineMessages({
-  empty: {
-    id: 'sources.index.empty',
-    defaultMessage: 'This group is not linked to any datasets.',
-  },
-});
+// const messages = defineMessages({
+//   empty: {
+//     id: 'sources.index.empty',
+//     defaultMessage: 'This group is not linked to any datasets.',
+//   },
+// });
 
 
 export class GroupScreen extends Component {
-  constructor(props) {
-    super(props);
-    this.updateQuery = this.updateQuery.bind(this);
-    this.getMoreResults = this.getMoreResults.bind(this);
-  }
-
   componentDidMount() {
     this.fetchIfNeeded();
   }
@@ -41,33 +35,15 @@ export class GroupScreen extends Component {
     this.fetchIfNeeded();
   }
 
-  getMoreResults() {
-    const { query, result } = this.props;
-    if (result && !result.isLoading && result.next && !result.isError) {
-      this.props.queryCollections({ query, result, next: result.next });
-    }
-  }
-
   fetchIfNeeded() {
-    const { query, result, groups } = this.props;
+    const { groups } = this.props;
     if (groups.shouldLoad) {
       this.props.fetchGroups();
     }
-    if (result.shouldLoad) {
-      this.props.queryCollections({ query });
-    }
-  }
-
-  updateQuery(newQuery) {
-    const { history, location } = this.props;
-    history.push({
-      pathname: location.pathname,
-      search: newQuery.toLocation(),
-    });
   }
 
   render() {
-    const { result, group, intl } = this.props;
+    const { query, group } = this.props;
     if (!group || !group.id) {
       return <LoadingScreen />;
     }
@@ -83,28 +59,9 @@ export class GroupScreen extends Component {
               />
             </p>
           </div>
-          {result.isError && (
-            <ErrorSection error={result.error} />
-          )}
-          {result.total === 0 && (
-            <ErrorSection
-              icon="shield"
-              title={intl.formatMessage(messages.empty)}
-            />
-          )}
-          <ul className="results">
-            {result.results !== undefined && result.results.map(
-              res => <CollectionListItem key={res.id} collection={res} />,
-            )}
-          </ul>
-          <Waypoint
-            onEnter={this.getMoreResults}
-            bottomOffset="-300px"
-            scrollableAncestor={window}
+          <CollectionIndex
+            query={query}
           />
-          {result.isLoading && (
-            <SectionLoading />
-          )}
         </Dashboard>
       </Screen>
     );
@@ -116,11 +73,11 @@ const mapStateToProps = (state, ownProps) => {
     .setFilter('team_id', groupId)
     .sortBy('count', 'desc')
     .limit(20);
+
   return {
     query,
     group: selectRole(state, groupId),
     groups: selectGroups(state),
-    result: selectCollectionsResult(state, query),
   };
 };
